Rename(lotto) : 로또 관련로직 패키지 이동
