+++ src/main/java/step3/lotto/BonusBall.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.lotto;
 
 public class BonusBall extends LottoNumber {
