+++ src/main/java/step3/lotto/LottoNumbers.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.lotto;
 
 import java.util.Arrays;
 import java.util.Collections;
