+++ src/main/java/step3/machine/Machine.java
@@ -1,6 +1,6 @@
 package step3.machine;
 
-import step3.LottoPapers;
+import step3.lotto.LottoPapers;
 import step3.Money;
 
 public interface Machine {
