+++ src/main/java/step3/winner/Winner.java
@@ -8,9 +8,9 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import step3.BonusBall;
-import step3.LottoNumbers;
-import step3.LottoPapers;
+import step3.lotto.BonusBall;
+import step3.lotto.LottoNumbers;
+import step3.lotto.LottoPapers;
 
 public class Winner {
